move intercom code into its own file
